Change complexity data type to double.
Unfortunately Charm++ doesn't allow a reduction on a size_t data type. The
regular int data type starts to overflow pretty quickly and is useless as a
complexity counter.

+++ spammpack-C/spamm-charm/src/multiplyelement.h
@@ -61,7 +61,7 @@ class MultiplyElement : public CBase_MultiplyElement
     void *CResult;
 
     /** The complexity of the last multiply(). */
-    int complexity;
+    size_t complexity;
 
 #ifndef PRUNE_CONVOLUTION
     /** A flag indicating whether this MultiplyElement is enabled or not (as a
